requirements.txt
pymupdf
selenium
rich
requests
pycryptodome
umsgpacktinydb
